refactor: Change component folder structure
created two primary folders:
 1. structural-components
 2. utility-components
