+++ src/app/app.component.spec.ts
@@ -1,12 +1,12 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
-import { AboutSectionComponent } from './about-section/about-section.component';
+import { AboutSectionComponent } from './structural-components/about-section/about-section.component';
 import { AppComponent } from './app.component';
-import { ColorSwitcherComponent } from './color-switcher/color-switcher.component';
-import { ContactSectionComponent } from './contact-section/contact-section.component';
-import { ExperienceSectionComponent } from './experience-section/experience-section.component';
-import { HeroSectionComponent } from './hero-section/hero-section.component';
-import { NavbarComponent } from './navbar/navbar.component';
+import { ColorSwitcherComponent } from './utility-components/color-switcher/color-switcher.component';
+import { ContactSectionComponent } from './structural-components/contact-section/contact-section.component';
+import { ExperienceSectionComponent } from './structural-components/experience-section/experience-section.component';
+import { HeroSectionComponent } from './structural-components/hero-section/hero-section.component';
+import { NavbarComponent } from './structural-components/navbar/navbar.component';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
