+++ src/app/structural-components/about-section/about-section.component.css
