+++ src/app/structural-components/footer/footer.component.spec.ts
