+++ src/app/structural-components/footer/footer.component.ts
