+++ src/app/structural-components/hero-section/hero-section.component.spec.ts
