+++ src/app/utility-components/color-switcher/color-switcher.component.css
