+++ src/app/utility-components/color-switcher/color-switcher.component.spec.ts
