+++ src/app/utility-components/plus-sign/plus-sign.component.css
